+++ bin/format_ppanggo_input_from_genome_selection.py
@@ -6,6 +6,7 @@ from pathlib import Path
 from typing import Dict, Set
 import gzip
 
+
 def parse_args(argv=None):
     """Define and immediately parse command line arguments."""
     parser = argparse.ArgumentParser(
@@ -65,13 +66,11 @@ def parse_args(argv=None):
     )
 
     parser.add_argument(
-        "--species",
-        type=str,
-        help="Specify species name",
-        default=None
+        "--species", type=str, help="Specify species name", default=None
     )
 
     return parser.parse_args(argv)
+
 
 def parse_genome_name_to_path_file(genome_name_to_path_file: Path) -> Dict[str, str]:
     """
@@ -91,9 +90,9 @@ def parse_genome_name_to_path_file(genome_name_to_path_file: Path) -> Dict[str, 
 
     return path_to_genome_name
 
+
 def get_path_of_reference_genomes(
-    reference_genome_names: Set[str],
-    path_to_genome_name: Dict[str, str]
+    reference_genome_names: Set[str], path_to_genome_name: Dict[str, str]
 ) -> Set[str]:
     """
     Retrieve the file paths of reference genomes that match the specified genome names.
@@ -103,10 +102,16 @@ def get_path_of_reference_genomes(
     :return: A set of file paths corresponding to the reference genomes.
     """
     # Identify reference genome names present in the provided genome name mapping
-    current_species_ref_genomes = set(path_to_genome_name.values()) & reference_genome_names
+    current_species_ref_genomes = (
+        set(path_to_genome_name.values()) & reference_genome_names
+    )
 
     # Collect file paths corresponding to the identified reference genome names
-    reference_genome_paths = {path for path, name in path_to_genome_name.items() if name in current_species_ref_genomes}
+    reference_genome_paths = {
+        path
+        for path, name in path_to_genome_name.items()
+        if name in current_species_ref_genomes
+    }
 
     return reference_genome_paths
 
@@ -150,11 +155,17 @@ def main(argv=None):
         with proper_open(args.reference_genomes, "rt") as fl:
             # Load reference genome names from the file
             reference_genomes = {line.rstrip() for line in fl if line}
-            logging.info(f"Loaded {len(reference_genomes)} reference genome names from {args.reference_genomes}.")
+            logging.info(
+                f"Loaded {len(reference_genomes)} reference genome names from {args.reference_genomes}."
+            )
 
         # Get paths for the specified reference genomes
-        reference_genome_paths = get_path_of_reference_genomes(reference_genomes, path_to_genome_name)
-        logging.info(f"Retrieved {len(reference_genome_paths)} reference genome paths matching the provided input genomes.")
+        reference_genome_paths = get_path_of_reference_genomes(
+            reference_genomes, path_to_genome_name
+        )
+        logging.info(
+            f"Retrieved {len(reference_genome_paths)} reference genome paths matching the provided input genomes."
+        )
 
         # Add the retrieved genome paths to the selected genomes set
         selection_count = len(selected_genomes)
@@ -186,14 +197,16 @@ def main(argv=None):
     logging.info(f"Writing summary selection to {args.summary_selection}.")
 
     final_selection = len(selected_genomes)
-    selection_no_ref =  final_selection - reference_genome_count
+    selection_no_ref = final_selection - reference_genome_count
     total_genomes = len(path_to_genome_name)
     discarded_genome_count = total_genomes - final_selection
 
     with open(args.summary_selection, "w") as fl:
 
-        fl.write('species\tselected_genomes\treference_genomes\tdiscarded_genomes\n')
-        fl.write(f'{args.species}\t{selection_no_ref}\t{reference_genome_count}\t{discarded_genome_count}\n')
+        fl.write("species\tselected_genomes\treference_genomes\tdiscarded_genomes\n")
+        fl.write(
+            f"{args.species}\t{selection_no_ref}\t{reference_genome_count}\t{discarded_genome_count}\n"
+        )
 
 
 if __name__ == "__main__":