+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### `Fixed`
 
+- Updated the storage of persistent families Mash sketch file to include only relative paths.
+
 ### `Dependencies`
 
 ### `Deprecated`