+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### `Added`
 
 - You can now associate external metadata with genomes using the `--genome_metadata` parameter. These metadata are stored in the pangenome file.
-- Ppanggolin process now produces tile plot to describe pangenome.
+- PPanGGOLiN process now produces tile plot to describe pangenome.
 - Add genome in pangenome stats summary at pangenome level.
 - Add proksee JSON map in the output.
 - Compute genomes fluidity for each pangenome.